tcp 통신 추가 및 수정
- 공제 중량 g으로 변경
- EditText SingleLine 으로 변경
- 박스 선택 시 박스 타이틀 색상 변경
- 기초 자료 품목 단가 수정 시 계량 작업에서도 같이 업데이트 되도록 수정
- 단가 null값이면 0으로 수정
- 계량 작업 화면 숫자 크기 변경
- 계량 내역 구분선 색상 변경

+++ app/src/main/java/com/example/wdgfarm_android/model/Weighing.java
@@ -38,7 +38,7 @@ public class Weighing {
 
     private int boxAccount;
     private float paletteWeight;
-    private float deductibleWeight;
+    private int deductibleWeight;
     private float realWeight;
 
     private String erpDate;
@@ -166,11 +166,11 @@ public class Weighing {
         this.paletteWeight = paletteWeight;
     }
 
-    public float getDeductibleWeight() {
+    public int getDeductibleWeight() {
         return deductibleWeight;
     }
 
-    public void setDeductibleWeight(float deductibleWeight) {
+    public void setDeductibleWeight(int deductibleWeight) {
         this.deductibleWeight = deductibleWeight;
     }
